refactor: changed tag entities and classes names
