+++ src/Application/Interfaces/Services/IInitiativeTagService.cs
@@ -1,12 +1,22 @@
 ﻿namespace IAVH.BioTablero.CM.Application.Interfaces.Services;
 
-using IAVH.BioTablero.CM.Application.DTOs.Initiatives;
+using System.Threading;
+using System.Threading.Tasks;
+
 using IAVH.BioTablero.CM.Application.Interfaces.General;
-using IAVH.BioTablero.CM.Core.Domain.Entities.Initiatives;
+using IAVH.BioTablero.CM.Application.Utils;
 
 /// <summary>
 /// Initiative Tag service interface.
 /// </summary>
-public interface IInitiativeTagService : IServiceRead<InitiativeTag, InitiativeTagDto, int>, IServiceAdd<InitiativeTagDto>, IServiceUpdate<InitiativeTagDto, int>, IServiceDelete<int>
+public interface IInitiativeTagService : IServiceDelete<int>
 {
+    /// <summary>
+    /// Add element.
+    /// </summary>
+    /// <param name="initiativeId">Initiative identifier.</param>
+    /// <param name="tagId">Initiative tag identifier.</param>
+    /// <param name="ct">Cancellation token.</param>
+    /// <returns>Process result.</returns>
+    Task<CustomWebResponse> Add(int initiativeId, int tagId, CancellationToken ct = default);
 }
