+++ src/Application/Services/Initiatives/InitiativeTagService.cs
@@ -1,16 +1,9 @@
 ﻿namespace IAVH.BioTablero.CM.Application.Services.Initiatives;
 
-using System;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
-using FluentValidation;
-
-using IAVH.BioTablero.CM.Application.DTOs.Initiatives;
-using IAVH.BioTablero.CM.Application.Interfaces.General;
 using IAVH.BioTablero.CM.Application.Interfaces.Services;
-using IAVH.BioTablero.CM.Application.Services.General;
 using IAVH.BioTablero.CM.Application.Specifications;
 using IAVH.BioTablero.CM.Application.Utils;
 using IAVH.BioTablero.CM.Core.Domain.Entities.Initiatives;
@@ -23,142 +16,87 @@ using static IAVH.BioTablero.CM.Core.Domain.Utils.Enums.LogEnums;
 /// <summary>
 /// Initiative Tag service.
 /// </summary>
-public class InitiativeTagService : ServiceRead<InitiativeTag, InitiativeTagDto, int, InitiativeTagSpec>, IInitiativeTagService
+public class InitiativeTagService : IInitiativeTagService
 {
-    private readonly IValidator<InitiativeTagDto> entityValidator;
+    private readonly IRepository<InitiativeTag> entityRepository;
     private readonly ILogger logger;
     private readonly IInitiativeRepository initiativeRepository;
+    private readonly IRepository<Tag> tagRepository;
 
     /// <summary>
     /// Constructor.
     /// </summary>
     /// <param name="entityRepository">Entity repository.</param>
-    /// <param name="mapper">Entity mapper.</param>
-    /// <param name="entityValidator">Entity validator.</param>
     /// <param name="logger">System logger.</param>
     /// <param name="initiativeRepository">Initiative repository.</param>
+    /// <param name="tagRepository">Initiative Tag repository.</param>
     public InitiativeTagService(
         IRepository<InitiativeTag> entityRepository,
-        IMapper<InitiativeTag, InitiativeTagDto> mapper,
-        IValidator<InitiativeTagDto> entityValidator,
         ILogger logger,
-        IInitiativeRepository initiativeRepository)
-        : base(entityRepository, mapper)
+        IInitiativeRepository initiativeRepository,
+        IRepository<Tag> tagRepository)
     {
-        this.entityValidator = entityValidator;
+        this.entityRepository = entityRepository;
         this.logger = logger;
         this.initiativeRepository = initiativeRepository;
+        this.tagRepository = tagRepository;
     }
 
     /// <summary>
     /// Add element.
     /// </summary>
-    /// <param name="entityData">Entity data.</param>
+    /// <param name="initiativeId">Initiative identifier.</param>
+    /// <param name="tagId">Initiative tag identifier.</param>
     /// <param name="ct">Cancellation token.</param>
     /// <returns>Process result.</returns>
-    public async Task<CustomWebResponse> Add(InitiativeTagDto entityData, CancellationToken ct = default)
+    public async Task<CustomWebResponse> Add(int initiativeId, int tagId, CancellationToken ct = default)
     {
-        // Validate data
-        entityData.Name = entityData.Name.Capitalize();
-        var validationResult = await entityValidator.ValidateAsync(entityData, options => options.IncludeRuleSets("default", "Create"), ct);
+        // Validate initiative
+        var initiativeExists = await initiativeRepository.AnyAsync(new InitiativeSpec(initiativeId), ct);
 
-        if (!validationResult.IsValid)
+        if (!initiativeExists)
         {
             return new CustomWebResponse(true)
             {
-                Message = "Validation errors",
-                ResponseBody = validationResult.Errors
-                    .Select(error => error.ErrorMessage),
+                Message = "Initiative not found",
+            };
+        }
+
+        // Validate initiative tag
+        var tagExists = await tagRepository.AnyAsync(new TagSpec(tagId), ct);
+
+        if (!tagExists)
+        {
+            return new CustomWebResponse(true)
+            {
+                Message = "Tag not found",
             };
         }
 
         // Validate duplicated entities
-        var hasDuplicatedEntities = await entityRepository.AnyAsync(new InitiativeTagSpec(entityData.Name), ct);
+        var hasDuplicatedEntities = await entityRepository.AnyAsync(InitiativeTagSpec.GetDuplicatesSpec(initiativeId, tagId), ct);
 
         if (hasDuplicatedEntities)
         {
             return new CustomWebResponse(true)
             {
-                Message = "Duplicated initiative tag",
+                Message = "Duplicated initiative tag relationship",
             };
         }
 
         // Build entity data
-        var entity = mapper.Map(entityData);
+        var entity = new InitiativeTag()
+        {
+            InitiativeId = initiativeId,
+            TagId = tagId,
+        };
 
         // Save data
         entity = await entityRepository.AddAsync(entity, ct);
 
-        entityData = mapper.Map(entity);
+        logger.AddLog(LogType.Create, "Added initiative tag relationship: {@entityData}", entity);
 
-        logger.AddLog(LogType.Create, "Added initiative tag: {@entityData}", entityData);
-
-        return new CustomWebResponse()
-        {
-            ResponseBody = entityData,
-        };
-    }
-
-    /// <summary>
-    /// Update element.
-    /// </summary>
-    /// <param name="id">Element identifier.</param>
-    /// <param name="entityData">Entity data.</param>
-    /// <param name="ct">Cancellation token.</param>
-    /// <returns>Process result.</returns>
-    public async Task<CustomWebResponse> Update(int id, InitiativeTagDto entityData, CancellationToken ct = default)
-    {
-        // Validate data
-        entityData.Name = entityData.Name.Capitalize();
-        var validationResult = await entityValidator.ValidateAsync(entityData, ct);
-
-        if (!validationResult.IsValid)
-        {
-            return new CustomWebResponse(true)
-            {
-                Message = "Validation errors",
-                ResponseBody = validationResult.Errors
-                    .Select(error => error.ErrorMessage),
-            };
-        }
-
-        // Validate entity
-        var entity = await entityRepository.GetByIdAsync(id, ct);
-
-        if (entity == null)
-        {
-            return new CustomWebResponse(true)
-            {
-                Message = "Not found",
-            };
-        }
-
-        // Validate duplicated entities
-        var hasDuplicatedEntities = await entityRepository.AnyAsync(new InitiativeTagSpec(id, entityData.Name), ct);
-
-        if (hasDuplicatedEntities)
-        {
-            return new CustomWebResponse(true)
-            {
-                Message = "Duplicated initiative tag",
-            };
-        }
-
-        // Update entity data
-        entity.Name = entityData.Name;
-        entity.Url = new Uri(entityData.Url);
-        entity.CategoryId = entityData.Category.Id;
-
-        await entityRepository.UpdateAsync(entity, ct);
-
-        entityData = mapper.Map(entity);
-
-        logger.AddLog(LogType.Update, "Updated initiative tag: {@entityData}", entityData);
-
-        return new CustomWebResponse()
-        {
-            ResponseBody = entityData,
-        };
+        return new CustomWebResponse();
     }
 
     /// <summary>
@@ -180,22 +118,9 @@ public class InitiativeTagService : ServiceRead<InitiativeTag, InitiativeTagDto,
             };
         }
 
-        // Validate relationships
-        var hasRelationships = await initiativeRepository.AnyAsync(InitiativeSpec.GetByTagSpec(id), ct);
-
-        if (hasRelationships)
-        {
-            return new CustomWebResponse(true)
-            {
-                Message = "The tag has relationships with one or more initiatives",
-            };
-        }
-
         await entityRepository.DeleteAsync(entity, ct);
 
-        var entityData = mapper.Map(entity);
-
-        logger.AddLog(LogType.Delete, "Deleted initiative tag: {@entityData}", entityData);
+        logger.AddLog(LogType.Delete, "Deleted initiative tag relationship: {@entityData}", entity);
 
         return new CustomWebResponse();
     }
