--- src/Core/Domain/Entities/Initiatives/InitiativeTagInitiative.cs
@@ -1,29 +1,0 @@
-﻿namespace IAVH.BioTablero.CM.Core.Domain.Entities.Initiatives;
-
-using IAVH.BioTablero.CM.Core.Interfaces.Entities;
-
-/// <summary>
-/// Initiative Tag Initiative entity.
-/// </summary>
-public class InitiativeTagInitiative : BaseEntity<int>, IAggregateRoot
-{
-    /// <summary>
-    /// Initiative identifier.
-    /// </summary>
-    public int InitiativeId { get; set; }
-
-    /// <summary>
-    /// Initiative Tag identifier.
-    /// </summary>
-    public int TagId { get; set; }
-
-    /// <summary>
-    /// Initiative relationship.
-    /// </summary>
-    public Initiative Initiative { get; set; }
-
-    /// <summary>
-    /// Initiative Tag relationship.
-    /// </summary>
-    public InitiativeTag Tag { get; set; }
-}
