+++ src/Infrastructure/Persistence/GeneralContext.cs
@@ -11,8 +11,8 @@ using IAVH.BioTablero.CM.Core.Domain.Entities.Logging;
 
 using Microsoft.EntityFrameworkCore;
 
-using InitiativeTagCategoryEnum = IAVH.BioTablero.CM.Core.Domain.Utils.Enums.InitiativesEnums.InitiativeTagCategory;
 using InitiativeUserLevelEnum = IAVH.BioTablero.CM.Core.Domain.Utils.Enums.InitiativesEnums.InitiativeUserLevel;
+using TagCategoryEnum = IAVH.BioTablero.CM.Core.Domain.Utils.Enums.InitiativesEnums.TagCategory;
 
 /// <summary>
 /// General database context.
@@ -83,17 +83,17 @@ public sealed class GeneralContext : DbContext
     /// <summary>
     /// Initiative tag initiatives DbSet.
     /// </summary>
-    public DbSet<InitiativeTagInitiative> InitiativeTagInitiatives { get; set; }
+    public DbSet<InitiativeTag> InitiativeTagInitiatives { get; set; }
 
     /// <summary>
     /// Initiative tags DbSet.
     /// </summary>
-    public DbSet<InitiativeTag> InitiativeTags { get; set; }
+    public DbSet<Tag> InitiativeTags { get; set; }
 
     /// <summary>
     /// Initiative tag categories DbSet.
     /// </summary>
-    public DbSet<InitiativeTagCategory> InitiativeTagCategories { get; set; }
+    public DbSet<TagCategory> InitiativeTagCategories { get; set; }
 
     #endregion
 
@@ -108,7 +108,7 @@ public sealed class GeneralContext : DbContext
 
         // Seeding data
         modelBuilder.Entity<InitiativeUserLevel>().HasData(GetDefaultInitiativeUserLevels());
-        modelBuilder.Entity<InitiativeTagCategory>().HasData(GetDefaultInitiativeTagCategories());
+        modelBuilder.Entity<TagCategory>().HasData(GetDefaultInitiativeTagCategories());
     }
 
     #region Seeding functions
@@ -132,11 +132,11 @@ public sealed class GeneralContext : DbContext
     /// Get default initative tag categories.
     /// </summary>
     /// <returns>Default initative tag categories list.</returns>
-    private static IEnumerable<InitiativeTagCategory> GetDefaultInitiativeTagCategories()
+    private static IEnumerable<TagCategory> GetDefaultInitiativeTagCategories()
     {
-        var enumData = Enum.GetValues(typeof(InitiativeTagCategoryEnum))
-            .Cast<InitiativeTagCategoryEnum>()
-            .Select(t => new InitiativeTagCategory() { Id = (int)t, Name = t.ToString() });
+        var enumData = Enum.GetValues(typeof(TagCategoryEnum))
+            .Cast<TagCategoryEnum>()
+            .Select(t => new TagCategory() { Id = (int)t, Name = t.ToString() });
 
         return enumData;
     }
