+++ src/WebApi/Config/DependencyRegistry/ConfigMappings.cs
@@ -29,7 +29,7 @@ public static class ConfigMappings
         services.AddSingleton<IMapper<InitiativeContact, InitiativeContactDto>, InitiativeContactMappings>();
         services.AddSingleton<IMapper<InitiativeLocation, InitiativeLocationDto>, InitiativeLocationMappings>();
         services.AddSingleton<IMapper<InitiativeUser, InitiativeUserDto>, InitiativeUserMappings>();
-        services.AddSingleton<IMapper<InitiativeTag, InitiativeTagDto>, InitiativeTagMappings>();
+        services.AddSingleton<IMapper<Tag, TagDto>, TagMappings>();
 
         return services;
     }
