+++ src/WebApi/Config/DependencyRegistry/ConfigValidators.cs
@@ -22,7 +22,7 @@ public static class ConfigValidators
         services.AddValidatorsFromAssemblyContaining<InitiativeContactValidator>();
         services.AddValidatorsFromAssemblyContaining<InitiativeUserValidator>();
         services.AddValidatorsFromAssemblyContaining<InitiativeLocationValidator>();
-        services.AddValidatorsFromAssemblyContaining<InitiativeTagValidator>();
+        services.AddValidatorsFromAssemblyContaining<TagValidator>();
 
         return services;
     }
