+++ src/WebApi/Config/DocsSetup/Examples/Initiative/TagResponseExample.cs
@@ -8,18 +8,19 @@ using Swashbuckle.AspNetCore.Filters;
 using static IAVH.BioTablero.CM.Core.Domain.Utils.Enums.InitiativesEnums;
 
 /// <summary>
-/// Initiative Tag add request example.
+/// Tag response example.
 /// </summary>
-public class InitiativeTagAddRequestExample : IExamplesProvider<InitiativeTagDto>
+public class TagResponseExample : IExamplesProvider<TagDto>
 {
     /// <summary>
     /// Get examples for entity.
     /// </summary>
     /// <returns>Entity examples.</returns>
-    public InitiativeTagDto GetExamples() => new()
+    public TagDto GetExamples() => new()
     {
+        Id = 0,
         Name = "Tag example",
         Url = "https://example.com/tag-data",
-        Category = new EnumEntityDto<InitiativeTagCategory>(InitiativeTagCategory.PoliticalContext),
+        Category = new EnumEntityDto<TagCategory>(TagCategory.PoliticalContext),
     };
 }
