+++ src/WebApi/Controllers/Rest/Initiatives/InitiativeTagController.cs
@@ -1,24 +1,15 @@
 ﻿namespace IAVH.BioTablero.CM.WebApi.Controllers.Rest.Initiatives;
 
-using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 
-using IAVH.BioTablero.CM.Application.DTOs.Initiatives;
 using IAVH.BioTablero.CM.Application.Interfaces.Services;
-using IAVH.BioTablero.CM.Core.Domain.Entities.Initiatives;
 using IAVH.BioTablero.CM.Core.Domain.Utils.Constants;
-using IAVH.BioTablero.CM.WebApi.Config.DocsSetup.Examples.General;
-using IAVH.BioTablero.CM.WebApi.Config.DocsSetup.Examples.Initiative;
 using IAVH.BioTablero.CM.WebApi.Interfaces;
 
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.OData.Query;
-using Microsoft.AspNetCore.OData.Routing.Controllers;
-
-using Swashbuckle.AspNetCore.Filters;
 
 /// <summary>
 /// Initiative Tag controller.
@@ -30,78 +21,23 @@ using Swashbuckle.AspNetCore.Filters;
 [Produces("application/json")]
 public class InitiativeTagController(
     IWebTools webTools,
-    IInitiativeTagService entityService) : ODataController
+    IInitiativeTagService entityService) : ControllerBase
 {
-    /// <summary>
-    /// Get entity.
-    /// </summary>
-    /// <param name="id">Entity identifier.</param>
-    /// <param name="ct">Cancellation token.</param>
-    /// <returns>Selected entity data.</returns>
-    [HttpGet("{id}")]
-    [ProducesResponseType(typeof(InitiativeTagDto), StatusCodes.Status200OK)]
-    [ProducesResponseType(StatusCodes.Status404NotFound)]
-    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-    [SwaggerResponseExample(StatusCodes.Status404NotFound, typeof(NotFoundResponseExample))]
-    [SwaggerResponseExample(StatusCodes.Status200OK, typeof(InitiativeTagResponseExample))]
-    public async Task<IActionResult> Get(int id, CancellationToken ct)
-    {
-        var response = await entityService.GetItem(id, ct);
-        return webTools.CustomResponse(response);
-    }
-
-    /// <summary>
-    /// Get entities (paginated).
-    /// </summary>
-    /// <param name="queryOptions">OData query options.</param>
-    /// <param name="ct">Cancellation token.</param>
-    /// <returns>Entities list from parameters.</returns>
-    [HttpGet]
-    [ProducesResponseType(typeof(InitiativeTagOdataResponseExample), StatusCodes.Status200OK)]
-    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-    [SwaggerResponseExample(StatusCodes.Status200OK, typeof(InitiativeTagOdataResponseExample))]
-    public async Task<IActionResult> Get(ODataQueryOptions<InitiativeTag> queryOptions, CancellationToken ct)
-    {
-        var response = await entityService.GetList(queryOptions, ct);
-        return webTools.CustomResponse(response);
-    }
-
     /// <summary>
     /// Add entity.
     /// </summary>
-    /// <param name="requestData">Request data.</param>
+    /// <param name="initiativeId">Initiative identifier.</param>
+    /// <param name="tagId">Tag identifier.</param>
     /// <param name="ct">Cancellation token.</param>
     /// <returns>Added entity data.</returns>
     [HttpPut]
     [Consumes("application/json")]
     [Authorize(Roles = IamConstants.RoleModuleAdmin)]
-    [SwaggerRequestExample(typeof(InitiativeTagDto), typeof(InitiativeTagAddRequestExample))]
-    [ProducesResponseType(typeof(InitiativeTagDto), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-    [SwaggerResponseExample(StatusCodes.Status200OK, typeof(InitiativeTagResponseExample))]
-    public async Task<IActionResult> Put([FromBody] InitiativeTagDto requestData, CancellationToken ct)
+    public async Task<IActionResult> Put(int initiativeId, int tagId, CancellationToken ct)
     {
-        var response = await entityService.Add(requestData, ct);
-        return webTools.CustomResponse(response);
-    }
-
-    /// <summary>
-    /// Edit entity.
-    /// </summary>
-    /// <param name="id">Entity identifier.</param>
-    /// <param name="requestData">Entity data.</param>
-    /// <param name="ct">Cancellation token.</param>
-    /// <returns>Updated entity data.</returns>
-    [HttpPost("{id}")]
-    [Consumes("application/json")]
-    [Authorize(Roles = IamConstants.RoleModuleAdmin)]
-    [SwaggerRequestExample(typeof(InitiativeTagDto), typeof(InitiativeTagResponseExample))]
-    [ProducesResponseType(typeof(InitiativeTagDto), StatusCodes.Status200OK)]
-    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-    [SwaggerResponseExample(StatusCodes.Status200OK, typeof(InitiativeTagResponseExample))]
-    public async Task<IActionResult> Post(int id, [FromBody] InitiativeTagDto requestData, CancellationToken ct)
-    {
-        var response = await entityService.Update(id, requestData, ct);
+        var response = await entityService.Add(initiativeId, tagId, ct);
         return webTools.CustomResponse(response);
     }
