--- src/WebApi/Controllers/Rest/Initiatives/InitiativeTagInitiativeController.cs
@@ -1,59 +1,0 @@
-﻿namespace IAVH.BioTablero.CM.WebApi.Controllers.Rest.Initiatives;
-
-using System.Threading;
-using System.Threading.Tasks;
-
-using IAVH.BioTablero.CM.Application.Interfaces.Services;
-using IAVH.BioTablero.CM.Core.Domain.Utils.Constants;
-using IAVH.BioTablero.CM.WebApi.Interfaces;
-
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-
-/// <summary>
-/// Initiative Tag Initiative controller.
-/// </summary>
-/// <param name="webTools">General web tools.</param>
-/// <param name="entityService">Entity service.</param>
-[ApiController]
-[Route("[controller]")]
-[Produces("application/json")]
-public class InitiativeTagInitiativeController(
-    IWebTools webTools,
-    IInitiativeTagInitiativeService entityService) : ControllerBase
-{
-    /// <summary>
-    /// Add entity.
-    /// </summary>
-    /// <param name="initiativeId">Initiative identifier.</param>
-    /// <param name="tagId">Initiative tag identifier.</param>
-    /// <param name="ct">Cancellation token.</param>
-    /// <returns>Added entity data.</returns>
-    [HttpPut]
-    [Consumes("application/json")]
-    [Authorize(Roles = IamConstants.RoleModuleAdmin)]
-    [ProducesResponseType(StatusCodes.Status200OK)]
-    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-    public async Task<IActionResult> Put(int initiativeId, int tagId, CancellationToken ct)
-    {
-        var response = await entityService.Add(initiativeId, tagId, ct);
-        return webTools.CustomResponse(response);
-    }
-
-    /// <summary>
-    /// Delete entity.
-    /// </summary>
-    /// <param name="id">Entity identifier.</param>
-    /// <param name="ct">Cancellation token.</param>
-    /// <returns>Process result.</returns>
-    [HttpDelete("{id}")]
-    [Authorize(Roles = IamConstants.RoleModuleAdmin)]
-    [ProducesResponseType(StatusCodes.Status200OK)]
-    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-    public async Task<IActionResult> Delete(int id, CancellationToken ct)
-    {
-        var response = await entityService.Delete(id, ct);
-        return webTools.CustomResponse(response);
-    }
-}
