+++ src/WebApi/Controllers/Rest/Initiatives/TagController.cs
@@ -1,0 +1,122 @@
+﻿namespace IAVH.BioTablero.CM.WebApi.Controllers.Rest.Initiatives;
+
+using System.Threading;
+using System.Threading.Tasks;
+
+using IAVH.BioTablero.CM.Application.DTOs.Initiatives;
+using IAVH.BioTablero.CM.Application.Interfaces.Services;
+using IAVH.BioTablero.CM.Core.Domain.Entities.Initiatives;
+using IAVH.BioTablero.CM.Core.Domain.Utils.Constants;
+using IAVH.BioTablero.CM.WebApi.Config.DocsSetup.Examples.General;
+using IAVH.BioTablero.CM.WebApi.Config.DocsSetup.Examples.Initiative;
+using IAVH.BioTablero.CM.WebApi.Interfaces;
+
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.OData.Query;
+using Microsoft.AspNetCore.OData.Routing.Controllers;
+
+using Swashbuckle.AspNetCore.Filters;
+
+/// <summary>
+/// Initiative Tag controller.
+/// </summary>
+/// <param name="webTools">General web tools.</param>
+/// <param name="entityService">Entity service.</param>
+[ApiController]
+[Route("[controller]")]
+[Produces("application/json")]
+public class TagController(
+    IWebTools webTools,
+    ITagService entityService) : ODataController
+{
+    /// <summary>
+    /// Get entity.
+    /// </summary>
+    /// <param name="id">Entity identifier.</param>
+    /// <param name="ct">Cancellation token.</param>
+    /// <returns>Selected entity data.</returns>
+    [HttpGet("{id}")]
+    [ProducesResponseType(typeof(TagDto), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+    [SwaggerResponseExample(StatusCodes.Status404NotFound, typeof(NotFoundResponseExample))]
+    [SwaggerResponseExample(StatusCodes.Status200OK, typeof(TagResponseExample))]
+    public async Task<IActionResult> Get(int id, CancellationToken ct)
+    {
+        var response = await entityService.GetItem(id, ct);
+        return webTools.CustomResponse(response);
+    }
+
+    /// <summary>
+    /// Get entities (paginated).
+    /// </summary>
+    /// <param name="queryOptions">OData query options.</param>
+    /// <param name="ct">Cancellation token.</param>
+    /// <returns>Entities list from parameters.</returns>
+    [HttpGet]
+    [ProducesResponseType(typeof(TagOdataResponseExample), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+    [SwaggerResponseExample(StatusCodes.Status200OK, typeof(TagOdataResponseExample))]
+    public async Task<IActionResult> Get(ODataQueryOptions<Tag> queryOptions, CancellationToken ct)
+    {
+        var response = await entityService.GetList(queryOptions, ct);
+        return webTools.CustomResponse(response);
+    }
+
+    /// <summary>
+    /// Add entity.
+    /// </summary>
+    /// <param name="requestData">Request data.</param>
+    /// <param name="ct">Cancellation token.</param>
+    /// <returns>Added entity data.</returns>
+    [HttpPut]
+    [Consumes("application/json")]
+    [Authorize(Roles = IamConstants.RoleModuleAdmin)]
+    [SwaggerRequestExample(typeof(TagDto), typeof(TagAddRequestExample))]
+    [ProducesResponseType(typeof(TagDto), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+    [SwaggerResponseExample(StatusCodes.Status200OK, typeof(TagResponseExample))]
+    public async Task<IActionResult> Put([FromBody] TagDto requestData, CancellationToken ct)
+    {
+        var response = await entityService.Add(requestData, ct);
+        return webTools.CustomResponse(response);
+    }
+
+    /// <summary>
+    /// Edit entity.
+    /// </summary>
+    /// <param name="id">Entity identifier.</param>
+    /// <param name="requestData">Entity data.</param>
+    /// <param name="ct">Cancellation token.</param>
+    /// <returns>Updated entity data.</returns>
+    [HttpPost("{id}")]
+    [Consumes("application/json")]
+    [Authorize(Roles = IamConstants.RoleModuleAdmin)]
+    [SwaggerRequestExample(typeof(TagDto), typeof(TagResponseExample))]
+    [ProducesResponseType(typeof(TagDto), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+    [SwaggerResponseExample(StatusCodes.Status200OK, typeof(TagResponseExample))]
+    public async Task<IActionResult> Post(int id, [FromBody] TagDto requestData, CancellationToken ct)
+    {
+        var response = await entityService.Update(id, requestData, ct);
+        return webTools.CustomResponse(response);
+    }
+
+    /// <summary>
+    /// Delete entity.
+    /// </summary>
+    /// <param name="id">Entity identifier.</param>
+    /// <param name="ct">Cancellation token.</param>
+    /// <returns>Process result.</returns>
+    [HttpDelete("{id}")]
+    [Authorize(Roles = IamConstants.RoleModuleAdmin)]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+    public async Task<IActionResult> Delete(int id, CancellationToken ct)
+    {
+        var response = await entityService.Delete(id, ct);
+        return webTools.CustomResponse(response);
+    }
+}
